Move `Rule` and `RuleWithNamedFields` into a `rule` module.
